requirements.txt
dash==2.15.0
dash-mantine-components==0.12.1
dash-iconify==0.1.2
pandas==2.2.1
psycopg2-binary==2.9.7